requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
attrs==23.2.0; python_version >= '3.7'
django==5.0.6; python_version >= '3.10'
djangorestframework==3.15.2; python_version >= '3.8'
drf-spectacular==0.27.2; python_version >= '3.7'
inflection==0.5.1; python_version >= '3.5'
jsonschema==4.22.0; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
markdown==3.6; python_version >= '3.8'
pyyaml==6.0.1; python_version >= '3.6'
referencing==0.35.1; python_version >= '3.8'
rpds-py==0.18.1; python_version >= '3.8'
sqlparse==0.5.0; python_version >= '3.8'
uritemplate==4.1.1; python_version >= '3.6'
